@@ -34,75 +35,78 @@
     "LICENSE"
   ],
   "dependencies": {
-    "@babel/runtime": "7.6.3",
+    "@babel/runtime": "7.8.7",
     "bcryptjs": "2.3.0",
     "body-parser": "1.19.0",
     "codemirror-graphql": "github:timsuchanek/codemirror-graphql#details-fix",
-    "commander": "3.0.2",
+    "commander": "5.0.0",
     "connect-flash": "0.1.1",
     "cookie-session": "2.0.0-beta.3",
     "copy-to-clipboard": "^3.2.0",
     "create-react-class": "15.6.3",
-    "csurf": "1.10.0",
+    "crypto-js": "^4.0.0",
+    "csurf": "1.11.0",
     "express": "4.17.1",
-    "graphql": "14.5.8",
+    "graphql": "14.6.0",
     "graphql-playground-react": "1.7.20",
     "history": "4.10.1",
     "immutable": "4.0.0-rc.9",
-    "immutable-devtools": "0.1.3",
-    "js-beautify": "1.10.2",
+    "immutable-devtools": "0.1.5",
+    "js-beautify": "1.11.0",
     "json-file-plus": "3.2.0",
     "package-json": "6.5.0",
-    "parse": "2.8.0",
-    "passport": "0.4.0",
+    "parse": "2.12.0",
+    "passport": "0.4.1",
     "passport-local": "1.0.0",
-    "prismjs": "1.17.0",
+    "prismjs": "1.20.0",
     "prop-types": "15.7.2",
-    "query-string": "6.8.3",
-    "react": "16.10.2",
-    "react-dnd": "9.4.0",
-    "react-dnd-html5-backend": "9.4.0",
-    "react-dom": "16.10.2",
-    "react-helmet": "5.2.1",
-    "react-redux": "5.1.1",
+    "query-string": "6.11.1",
+    "react": "16.13.1",
+    "react-ace": "^8.0.0",
+    "react-dnd": "10.0.2",
+    "react-dnd-html5-backend": "10.0.2",
+    "react-dom": "16.13.1",
+    "react-helmet": "6.0.0",
+    "react-json-view": "^1.19.1",
+    "react-redux": "5.1.2",
     "react-router": "5.1.2",
     "react-router-dom": "5.1.2",
-    "regenerator-runtime": "0.13.3",
-    "semver": "6.3.0"
+    "regenerator-runtime": "0.13.5",
+    "semver": "7.3.2"
   },
   "devDependencies": {
-    "@babel/core": "7.6.4",
-    "@babel/plugin-proposal-decorators": "7.6.0",
-    "@babel/plugin-transform-regenerator": "7.4.5",
-    "@babel/plugin-transform-runtime": "7.6.2",
-    "@babel/preset-env": "7.6.3",
-    "@babel/preset-react": "7.6.3",
-    "babel-eslint": "10.0.3",
-    "babel-loader": "8.0.6",
+    "@babel/core": "7.8.7",
+    "@babel/plugin-proposal-decorators": "7.8.3",
+    "@babel/plugin-transform-regenerator": "7.8.7",
+    "@babel/plugin-transform-runtime": "7.8.3",
+    "@babel/preset-env": "7.8.7",
+    "@babel/preset-react": "7.8.3",
+    "babel-eslint": "10.1.0",
+    "babel-loader": "8.1.0",
     "babel-plugin-transform-object-rest-spread": "6.26.0",
-    "core-js": "3.3.2",
-    "cross-env": "6.0.3",
-    "css-loader": "3.2.0",
-    "eslint": "6.5.1",
-    "eslint-plugin-jest": "22.19.0",
-    "eslint-plugin-react": "7.16.0",
-    "file-loader": "4.2.0",
-    "http-server": "0.11.1",
+    "core-js": "3.6.5",
+    "cross-env": "7.0.2",
+    "css-loader": "3.5.2",
+    "eslint": "6.8.0",
+    "eslint-plugin-jest": "23.8.2",
+    "eslint-plugin-react": "7.19.0",
+    "file-loader": "6.0.0",
+    "http-server": "0.12.0",
     "jest": "24.8.0",
-    "marked": "0.7.0",
-    "node-sass": "4.12.0",
+    "marked": "0.8.2",
+    "node-sass": "^4.13.1",
     "null-loader": "^3.0.0",
     "parse-server": "github:parseplatform/parse-server#master",
-    "path-to-regexp": "3.1.0",
-    "puppeteer": "1.20.0",
-    "react-test-renderer": "16.10.2",
-    "request": "2.87.0",
-    "request-promise": "4.2.4",
+    "path-to-regexp": "3.2.0",
+    "puppeteer": "3.0.0",
+    "react-test-renderer": "16.13.1",
+    "request": "2.88.2",
+    "request-promise": "4.2.5",
     "sass-loader": "8.0.0",
-    "style-loader": "1.0.0",
+    "style-loader": "1.1.2",
     "svg-prep": "1.0.4",
-    "webpack": "4.41.2",
-    "webpack-cli": "3.3.9"
+    "webpack": "4.42.1",
+    "webpack-cli": "3.3.10"
   },
   "scripts": {
     "dev": "node ./Parse-Dashboard/index.js & webpack --config webpack/build.config.js --devtool eval-source-map --progress --watch",
