@@ -19,7 +19,7 @@ export default class BrowserRow extends Component {
   }
 
   render() {
-    const { className, columns, currentCol, isUnique, obj, onPointerClick, order, readOnlyFields, row, rowWidth, selection, selectRow, setCopyableValue, setCurrent, setEditing, setRelation } = this.props;
+    const { className, columns, currentCol, isUnique, obj, onPointerClick, order, readOnlyFields, row, rowWidth, selection, selectRow, setCopyableValue, setCurrent, setEditing, setRelation, onEditSelectedRow, setContextMenu, onFilterChange } = this.props;
     let attributes = obj.attributes;
     return (
       <div className={styles.tableRow} style={{ minWidth: rowWidth }}>
@@ -61,6 +61,11 @@ export default class BrowserRow extends Component {
           return (
             <BrowserCell
               key={name}
+              schema={this.props.schema}
+              simplifiedSchema={this.props.simplifiedSchema}
+              filters={this.props.filters}
+              className={className}
+              field={name}
               row={row}
               col={j}
               type={type}
@@ -70,10 +75,14 @@ export default class BrowserRow extends Component {
               onSelect={setCurrent}
               onEditChange={setEditing}
               onPointerClick={onPointerClick}
+              onFilterChange={onFilterChange}
               setRelation={setRelation}
+              objectId={obj.id}
               value={attr}
               hidden={hidden}
-              setCopyableValue={setCopyableValue} />
+              setCopyableValue={setCopyableValue}
+              setContextMenu={setContextMenu}
+              onEditSelectedRow={onEditSelectedRow} />
           );
         })}
       </div>
