@@ -20,6 +20,7 @@ export default class FileEditor extends React.Component {
 
     this.checkExternalClick = this.checkExternalClick.bind(this);
     this.handleKey = this.handleKey.bind(this);
+    this.removeFile = this.removeFile.bind(this);
   }
 
   componentDidMount() {
@@ -33,24 +34,31 @@ export default class FileEditor extends React.Component {
   }
 
   checkExternalClick(e) {
-    if (!hasAncestor(e.target, this.refs.input)) {
-      this.props.onCommit(this.state.value);
+    const { onCancel } = this.props;
+    if (!hasAncestor(e.target, this.refs.input) && onCancel) {
+      onCancel();
     }
   }
 
   handleKey(e) {
-    if (e.keyCode === 13) {
-      this.props.onCommit(this.state.value);
+    const { onCancel } = this.props;
+    if (e.keyCode === 13 && onCancel) {
+      onCancel();
     }
   }
 
-  getBase64(file){
+  getBase64(file) {
     return new Promise((resolve, reject) => {
       const reader = new FileReader();
       reader.readAsDataURL(file);
       reader.onload = () => resolve(reader.result);
       reader.onerror = error => reject(error);
     });
+  }
+
+  removeFile() {
+    this.refs.fileInput.value = '';
+    this.props.onCommit(undefined);
   }
 
   async handleChange(e) {
@@ -67,10 +75,10 @@ export default class FileEditor extends React.Component {
       <div ref='input' style={{ minWidth: this.props.width }} className={styles.editor}>
         {file && file.url() ? <a href={file.url()} target='_blank' role='button' className={styles.download}>Download</a> : null}
         <a className={styles.upload}>
-          <input type='file' onChange={this.handleChange.bind(this)} />
+          <input ref='fileInput' type='file' onChange={this.handleChange.bind(this)} />
           <span>{file ? 'Replace file' : 'Upload file'}</span>
         </a>
-        {file ? <a href='javascript:;' role='button' className={styles.delete} onClick={() => this.props.onCommit(undefined)}>Delete</a> : null}
+        {file ? <a href='javascript:;' role='button' className={styles.delete} onClick={this.removeFile}>Delete</a> : null}
       </div>
     );
   }
