@@ -5,12 +5,11 @@
  * This source code is licensed under the license found in the LICENSE file in
  * the root directory of this source tree.
  */
-import React from 'react'
-import CategoryList from 'components/CategoryList/CategoryList.react'
-import DashboardView from 'dashboard/DashboardView.react'
+import React from 'react';
+import CategoryList from 'components/CategoryList/CategoryList.react';
+import DashboardView from 'dashboard/DashboardView.react';
 
 export default class ApiConsole extends DashboardView {
-
   constructor() {
     super();
     this.section = 'Core';
@@ -18,21 +17,23 @@ export default class ApiConsole extends DashboardView {
   }
 
   renderSidebar() {
-    const { path } = this.props.match
-    const current = path.substr(path.lastIndexOf('/') + 1, path.length - 1)
+    const { path } = this.props.match;
+    const current = path.substr(path.lastIndexOf('/') + 1, path.length - 1);
     return (
-      <CategoryList current={current} linkPrefix={'api_console/'} categories={[
-        { name: 'REST Console', id: 'rest' },
-        { name: 'GraphQL Console', id: 'graphql' }
-      ]} />
-    )
+      <CategoryList
+        current={current}
+        linkPrefix={'api_console/'}
+        categories={[
+          { name: 'REST Console', id: 'rest' },
+          { name: 'GraphQL Console', id: 'graphql' },
+          { name: 'JS Console', id: 'js_console' }
+        ]}
+      />
+    );
   }
 
   renderContent() {
     const child = React.Children.only(this.props.children);
-    return React.cloneElement(
-      child,
-      { ...child.props }
-    )
+    return React.cloneElement(child, { ...child.props });
   }
 }
