@@ -19,6 +19,7 @@ import ExportDialog                       from 'dashboard/Data/Browser/ExportDia
 import AttachRowsDialog                   from 'dashboard/Data/Browser/AttachRowsDialog.react';
 import AttachSelectedRowsDialog           from 'dashboard/Data/Browser/AttachSelectedRowsDialog.react';
 import CloneSelectedRowsDialog            from 'dashboard/Data/Browser/CloneSelectedRowsDialog.react';
+import EditRowDialog                      from 'dashboard/Data/Browser/EditRowDialog.react';
 import history                            from 'dashboard/history';
 import { List, Map }                      from 'immutable';
 import Notification                       from 'dashboard/Data/Browser/Notification.react';
@@ -56,6 +57,7 @@ class Browser extends DashboardView {
       showDropClassDialog: false,
       showExportDialog: false,
       showAttachRowsDialog: false,
+      showEditRowDialog: false,
       rowsToDelete: null,
 
       relation: null,
@@ -109,11 +111,16 @@ class Browser extends DashboardView {
     this.setRelation = this.setRelation.bind(this);
     this.showAddColumn = this.showAddColumn.bind(this);
     this.addRow = this.addRow.bind(this);
+    this.addRowWithModal = this.addRowWithModal.bind(this);
     this.showCreateClass = this.showCreateClass.bind(this);
     this.createClass = this.createClass.bind(this);
     this.addColumn = this.addColumn.bind(this);
     this.removeColumn = this.removeColumn.bind(this);
     this.showNote = this.showNote.bind(this);
+    this.showEditRowDialog = this.showEditRowDialog.bind(this);
+    this.closeEditRowDialog = this.closeEditRowDialog.bind(this);
+    this.handleShowAcl = this.handleShowAcl.bind(this);
+    this.onDialogToggle = this.onDialogToggle.bind(this);
   }
 
   componentWillMount() {
@@ -287,6 +294,12 @@ class Browser extends DashboardView {
     }
   }
 
+  addRowWithModal() {
+    this.addRow();
+    this.selectRow(undefined, true);
+    this.showEditRowDialog();
+  }
+
   removeColumn(name) {
     let payload = {
       className: this.props.params.className,
@@ -405,7 +418,7 @@ class Browser extends DashboardView {
     let className = this.props.params.className;
     let source = this.state.relation || className;
     let query = queryFromFilters(source, this.state.filters);
-    if (this.state.ordering !== 'createdAt') {
+    if (this.state.ordering !== '-createdAt') {
       // Construct complex pagination query
       let equalityQuery = queryFromFilters(source, this.state.filters);
       let field = this.state.ordering;
@@ -421,8 +434,12 @@ class Browser extends DashboardView {
       } else {
         query.greaterThan(field, comp);
       }
-      equalityQuery.equalTo(field, comp);
-      equalityQuery.lessThan('createdAt', this.state.data[this.state.data.length - 1].get('createdAt'));
+      if (field === 'createdAt') {
+        equalityQuery.greaterThan('createdAt', this.state.data[this.state.data.length - 1].get('createdAt'));
+      } else {
+        equalityQuery.lessThan('createdAt', this.state.data[this.state.data.length - 1].get('createdAt'));
+        equalityQuery.equalTo(field, comp);
+      }
       query = Parse.Query.or(query, equalityQuery);
       if (ascending) {
         query.ascending(this.state.ordering);
@@ -431,8 +448,8 @@ class Browser extends DashboardView {
       }
     } else {
       query.lessThan('createdAt', this.state.data[this.state.data.length - 1].get('createdAt'));
+      query.addDescending('createdAt');
     }
-    query.addDescending('createdAt');
     query.limit(MAX_ROWS_FETCHED);
 
     query.find({ useMasterKey: true }).then((nextPage) => {
@@ -492,9 +509,9 @@ class Browser extends DashboardView {
     });
   }
 
-  handlePointerClick({ className, id }) {
+  handlePointerClick({ className, id, field = 'objectId' }) {
     let filters = JSON.stringify([{
-        field: 'objectId',
+        field,
         constraint: 'eq',
         compareTo: id
     }]);
@@ -526,8 +543,8 @@ class Browser extends DashboardView {
       obj.set(attr, value);
     }
     obj.save(null, { useMasterKey: true }).then((objectSaved) => {
-      const createdOrUpdated = isNewObject ? "created" : "updated";
-      let msg = objectSaved.className + " with id '" + objectSaved.id + "' " + createdOrUpdated;
+      const createdOrUpdated = isNewObject ? 'created' : 'updated';
+      let msg = objectSaved.className + ' with id \'' + objectSaved.id + '\' ' + createdOrUpdated;
       this.showNote(msg, false);
 
       const state = { data: this.state.data };
@@ -634,9 +651,9 @@ class Browser extends DashboardView {
           let deletedNote;
 
           if (toDeleteObjectIds.length == 1) {
-            deletedNote = className + " with id '" + toDeleteObjectIds[0] + "' deleted";
+            deletedNote = className + ' with id \'' + toDeleteObjectIds[0] + '\' deleted';
           } else {
-            deletedNote = toDeleteObjectIds.length + " " + className + " objects deleted";
+            deletedNote = toDeleteObjectIds.length + ' ' + className + ' objects deleted';
           }
 
           this.showNote(deletedNote, false);
@@ -660,17 +677,17 @@ class Browser extends DashboardView {
 
           if (error.code === Parse.Error.AGGREGATE_ERROR) {
             if (error.errors.length == 1) {
-              errorDeletingNote = "Error deleting " + className + " with id '" + error.errors[0].object.id + "'";
+              errorDeletingNote = 'Error deleting ' + className + ' with id \'' + error.errors[0].object.id + '\'';
             } else if (error.errors.length < toDeleteObjectIds.length) {
-              errorDeletingNote = "Error deleting " + error.errors.length + " out of " + toDeleteObjectIds.length + " " + className + " objects";
+              errorDeletingNote = 'Error deleting ' + error.errors.length + ' out of ' + toDeleteObjectIds.length + ' ' + className + ' objects';
             } else {
-              errorDeletingNote = "Error deleting all " + error.errors.length + " " + className + " objects";
+              errorDeletingNote = 'Error deleting all ' + error.errors.length + ' ' + className + ' objects';
             }
           } else {
             if (toDeleteObjectIds.length == 1) {
-              errorDeletingNote = "Error deleting " + className + " with id '" + toDeleteObjectIds[0] + "'";
+              errorDeletingNote = 'Error deleting ' + className + ' with id \'' + toDeleteObjectIds[0] + '\'';
             } else {
-              errorDeletingNote = "Error deleting " + toDeleteObjectIds.length + " " + className + " objects";
+              errorDeletingNote = 'Error deleting ' + toDeleteObjectIds.length + ' ' + className + ' objects';
             }
           }
 
@@ -701,7 +718,9 @@ class Browser extends DashboardView {
       this.state.rowsToDelete ||
       this.state.showAttachRowsDialog ||
       this.state.showAttachSelectedRowsDialog ||
-      this.state.showCloneSelectedRowsDialog
+      this.state.showCloneSelectedRowsDialog ||
+      this.state.showEditRowDialog ||
+      this.state.showPermissionsDialog
     );
   }
 
@@ -764,10 +783,10 @@ class Browser extends DashboardView {
     });
   }
 
-  async confirmAttachSelectedRows(className, targetObjectId, relationName, objectIds) {
+  async confirmAttachSelectedRows(className, targetObjectId, relationName, objectIds, targetClassName) {
     const parentQuery = new Parse.Query(className);
     const parent = await parentQuery.get(targetObjectId, { useMasterKey: true });
-    const query = new Parse.Query(this.props.params.className);
+    const query = new Parse.Query(targetClassName || this.props.params.className);
     query.containedIn('objectId', objectIds);
     const objects = await query.find({ useMasterKey: true });
     parent.relation(relationName).add(objects);
@@ -891,6 +910,34 @@ class Browser extends DashboardView {
     }, 3500);
   }
 
+  showEditRowDialog(selectRow, objectId) {
+    // objectId is optional param which is used for doubleClick event on objectId BrowserCell
+    if (selectRow) {
+      // remove all selected rows and select doubleClicked row
+      this.setState({ selection: {} });
+      this.selectRow(objectId, true);
+    }
+    this.setState({
+      showEditRowDialog: true,
+    });
+  }
+
+  closeEditRowDialog() {
+    this.setState({
+      showEditRowDialog: false,
+    });
+  }
+
+  handleShowAcl(row, col){
+    this.refs.dataBrowser.setEditing(true);
+    this.refs.dataBrowser.setCurrent({ row, col });
+  }
+
+  // skips key controls handling when dialog is opened
+  onDialogToggle(opened){
+    this.setState({showPermissionsDialog: opened});
+  }
+  
   renderContent() {
     let browser = null;
     let className = this.props.params.className;
@@ -941,6 +988,7 @@ class Browser extends DashboardView {
         }
         browser = (
           <DataBrowser
+            ref='dataBrowser'
             isUnique={this.state.isUnique}
             uniqueField={this.state.uniqueField}
             count={count}
@@ -957,6 +1005,8 @@ class Browser extends DashboardView {
             onAttachRows={this.showAttachRowsDialog}
             onAttachSelectedRows={this.showAttachSelectedRowsDialog}
             onCloneSelectedRows={this.showCloneSelectedRowsDialog}
+            onEditSelectedRow={this.showEditRowDialog}
+            onEditPermissions={this.onDialogToggle}
 
             columns={columns}
             className={className}
@@ -975,6 +1025,7 @@ class Browser extends DashboardView {
             setRelation={this.setRelation}
             onAddColumn={this.showAddColumn}
             onAddRow={this.addRow}
+            onAddRowWithModal={this.addRowWithModal}
             onAddClass={this.showCreateClass}
             showNote={this.showNote} />
         );
@@ -1064,6 +1115,66 @@ class Browser extends DashboardView {
           onConfirm={this.confirmCloneSelectedRows}
         />
       );
+    } else if (this.state.showEditRowDialog) {
+      const classColumns = this.getClassColumns(className, false);
+      // create object with classColumns as property keys needed for ColumnPreferences.getOrder function
+      const columnsObject = {};
+      classColumns.forEach((column) => {
+        columnsObject[column.name] = column
+      });
+      // get ordered list of class columns
+      const columns = ColumnPreferences.getOrder(
+        columnsObject,
+        this.context.currentApp.applicationId,
+        className
+      );
+      // extend columns with their type and targetClass properties
+      columns.forEach(column => {
+        const { type, targetClass } = columnsObject[column.name];
+        column.type = type;
+        column.targetClass = targetClass;
+      });
+
+      const { data, selection, newObject } = this.state;
+      // at this moment only one row must be selected, so take the first and only one
+      const selectedKey = Object.keys(selection)[0];
+      // if selectedKey is string "undefined" => new row column 'objectId' was clicked
+      let selectedId = selectedKey === 'undefined' ? undefined : selectedKey;
+      // if selectedId is undefined and newObject is null it means new row was just saved and added to data array
+      const isJustSavedObject = selectedId === undefined && newObject === null;
+      // if new object just saved, remove new row selection and select new added row
+      if (isJustSavedObject) {
+        selectedId = data[0].id;
+        this.setState({ selection: {} });
+        this.selectRow(selectedId, true);
+      }
+
+      const row = data.findIndex(d => d.id === selectedId);
+
+      const attributes = selectedId
+        ? data[row].attributes
+        : newObject.attributes;
+
+      const selectedObject = {
+        row: row,
+        id: selectedId,
+        ...attributes
+      };
+
+      extras = (
+        <EditRowDialog
+          className={className}
+          columns={columns}
+          selectedObject={selectedObject}
+          handlePointerClick={this.handlePointerClick}
+          setRelation={this.setRelation}
+          handleShowAcl={this.handleShowAcl}
+          onClose={this.closeEditRowDialog}
+          updateRow={this.updateRow}
+          confirmAttachSelectedRows={this.confirmAttachSelectedRows}
+          schema={this.props.schema}
+        />
+      )
     }
 
     let notification = null;
