@@ -5,18 +5,18 @@
  * This source code is licensed under the license found in the LICENSE file in
  * the root directory of this source tree.
  */
-import BrowserFilter  from 'components/BrowserFilter/BrowserFilter.react';
-import BrowserMenu    from 'components/BrowserMenu/BrowserMenu.react';
-import ColumnsConfiguration
-                      from 'components/ColumnsConfiguration/ColumnsConfiguration.react';
-import Icon           from 'components/Icon/Icon.react';
-import MenuItem       from 'components/BrowserMenu/MenuItem.react';
-import prettyNumber   from 'lib/prettyNumber';
-import React          from 'react';
-import SecurityDialog from 'dashboard/Data/Browser/SecurityDialog.react';
-import Separator      from 'components/BrowserMenu/Separator.react';
-import styles         from 'dashboard/Data/Browser/Browser.scss';
-import Toolbar        from 'components/Toolbar/Toolbar.react';
+import BrowserFilter        from 'components/BrowserFilter/BrowserFilter.react';
+import BrowserMenu          from 'components/BrowserMenu/BrowserMenu.react';
+import Icon                 from 'components/Icon/Icon.react';
+import MenuItem             from 'components/BrowserMenu/MenuItem.react';
+import prettyNumber         from 'lib/prettyNumber';
+import React, { useRef }    from 'react';
+import Separator            from 'components/BrowserMenu/Separator.react';
+import styles               from 'dashboard/Data/Browser/Browser.scss';
+import Toolbar              from 'components/Toolbar/Toolbar.react';
+import SecurityDialog       from 'dashboard/Data/Browser/SecurityDialog.react';
+import ColumnsConfiguration  from 'components/ColumnsConfiguration/ColumnsConfiguration.react'
+import SecureFieldsDialog   from 'dashboard/Data/Browser/SecureFieldsDialog.react';
 
 let BrowserToolbar = ({
   className,
@@ -31,7 +31,9 @@ let BrowserToolbar = ({
   onFilterChange,
   onAddColumn,
   onAddRow,
+  onAddRowWithModal,
   onAddClass,
+  onEditSelectedRow,
   onAttachRows,
   onAttachSelectedRows,
   onCloneSelectedRows,
@@ -41,6 +43,7 @@ let BrowserToolbar = ({
   onDropClass,
   onChangeCLP,
   onRefresh,
+  onEditPermissions,
   hidePerms,
   isUnique,
   uniqueField,
@@ -98,12 +101,19 @@ let BrowserToolbar = ({
         />
       </BrowserMenu>
     );
-  } else {
+  } else if (onAddRow) {
     menu = (
       <BrowserMenu title='Edit' icon='edit-solid' disabled={isUnique}>
         <MenuItem text='Add a row' onClick={onAddRow} />
+        <MenuItem text='Add a row with modal' onClick={onAddRowWithModal} />
         {enableColumnManipulation ? <MenuItem text='Add a column' onClick={onAddColumn} /> : <noscript />}
         {enableClassManipulation ? <MenuItem text='Add a class' onClick={onAddClass} /> : <noscript />}
+        <Separator />
+        <MenuItem
+          disabled={selectionLength !== 1}
+          text={'Edit this row with modal'}
+          onClick={onEditSelectedRow}
+        />
         <Separator />
         <MenuItem
           disabled={!selectionLength}
@@ -143,6 +153,7 @@ let BrowserToolbar = ({
     onClick = null;
   }
 
+  const columns = {};
   const userPointers = [];
   const schemaSimplifiedData = {};
   const classSchema = schema.data.get('classes').get(classNameForEditors);
@@ -153,14 +164,22 @@ let BrowserToolbar = ({
         targetClass,
       };
 
+      columns[col] = { type, targetClass };
+
       if (col === 'objectId' || isUnique && col !== uniqueField) {
         return;
       }
-      if (targetClass === '_User') {
+      if ((type ==='Pointer' && targetClass === '_User') || type === 'Array' ) {
         userPointers.push(col);
       }
     });
   }
+
+  let clpDialogRef = useRef(null);
+  let protectedDialogRef = useRef(null);
+
+  const showCLP = ()=> clpDialogRef.current.handleOpen();
+  const showProtected = () => protectedDialogRef.current.handleOpen();
 
   return (
     <Toolbar
@@ -170,18 +189,21 @@ let BrowserToolbar = ({
       subsection={subsection}
       details={details.join(' \u2022 ')}
     >
-      <a className={classes.join(' ')} onClick={onClick}>
-        <Icon name='plus-solid' width={14} height={14} />
-        <span>Add Row</span>
-      </a>
-      <div className={styles.toolbarSeparator} />
+      {onAddRow && (
+        <a className={classes.join(' ')} onClick={onClick}>
+          <Icon name="plus-solid" width={14} height={14} />
+          <span>Add Row</span>
+        </a>
+      )}
+      {onAddRow && <div className={styles.toolbarSeparator} />}
       <ColumnsConfiguration
         handleColumnsOrder={handleColumnsOrder}
         handleColumnDragDrop={handleColumnDragDrop}
-        order={order} />
+        order={order}
+      />
       <div className={styles.toolbarSeparator} />
       <a className={styles.toolbarButton} onClick={onRefresh}>
-        <Icon name='refresh-solid' width={14} height={14} />
+        <Icon name="refresh-solid" width={14} height={14} />
         <span>Refresh</span>
       </a>
       <div className={styles.toolbarSeparator} />
@@ -190,16 +212,56 @@ let BrowserToolbar = ({
         schema={schemaSimplifiedData}
         filters={filters}
         onChange={onFilterChange}
-        className={classNameForEditors} />
-      <div className={styles.toolbarSeparator} />
-      {enableSecurityDialog ? <SecurityDialog
-        setCurrent={setCurrent}
+        className={classNameForEditors}
+        blacklistedFilters={onAddRow ? [] : ['unique']}
+      />
+      {onAddRow && <div className={styles.toolbarSeparator} />}
+      {perms && enableSecurityDialog ? (
+        <SecurityDialog
+          ref={clpDialogRef}
+          disabled={!!relation || !!isUnique}
+          perms={perms}
+          columns={columns}
+          className={classNameForEditors}
+          onChangeCLP={onChangeCLP}
+          userPointers={userPointers}
+          title="ClassLevelPermissions"
+          icon="locked-solid"
+          onEditPermissions={onEditPermissions}
+        />
+      ) : (
+        <noscript />
+      )}
+      <SecureFieldsDialog
+        ref={protectedDialogRef}
+        columns={columns}
         disabled={!!relation || !!isUnique}
         perms={perms}
         className={classNameForEditors}
         onChangeCLP={onChangeCLP}
-        userPointers={userPointers} /> : <noscript />}
-      {enableSecurityDialog ? <div className={styles.toolbarSeparator} /> : <noscript/>}
+        userPointers={userPointers}
+        title='ProtectedFields'
+        icon='locked-solid'
+        onEditPermissions={onEditPermissions}
+      />
+      {enableSecurityDialog ? (
+        <BrowserMenu
+          setCurrent={setCurrent}
+          title="Security"
+          icon="locked-solid"
+          disabled={!!relation || !!isUnique}
+        >
+          <MenuItem text={'Class Level Permissions'} onClick={showCLP} />
+          <MenuItem text={'Protected Fields'} onClick={showProtected} />
+        </BrowserMenu>
+      ) : (
+        <noscript />
+      )}
+      {enableSecurityDialog ? (
+        <div className={styles.toolbarSeparator} />
+      ) : (
+        <noscript />
+      )}
       {menu}
     </Toolbar>
   );
