@@ -47,7 +47,7 @@ export default class ExportDialog extends React.Component {
         icon='down-outline'
         iconSize={40}
         title='Export this class'
-        subtitle={`We'll send you an email when your data is ready.`}
+        subtitle={'We\'ll send you an email when your data is ready.'}
         confirmText='Export'
         cancelText='Cancel'
         disabled={this.state.progress === undefined || inProgress}
