@@ -19,7 +19,7 @@ import TextInput       from 'components/TextInput/TextInput.react';
 import Toggle          from 'components/Toggle/Toggle.react';
 import validateNumeric from 'lib/validateNumeric';
 import styles          from 'dashboard/Data/Browser/Browser.scss';
-import semver          from 'semver';
+import semver          from 'semver/preload.js';
 
 const PARAM_TYPES = [
   'Boolean',
