@@ -61,7 +61,7 @@ export default class HostingSettings extends DashboardView {
 				labelWidth={62}
 				label={<Label
 					text="Host name"
-					description={"A host with a subdomain CNAME entry set to " + (fields.subdomain_name || '[your subdomain name]') + ".parseapp.com"}
+					description={'A host with a subdomain CNAME entry set to ' + (fields.subdomain_name || '[your subdomain name]') + '.parseapp.com'}
 				/>}
 				input={<TextInput
 					value={fields.host_name}
@@ -154,7 +154,7 @@ export default class HostingSettings extends DashboardView {
 				labelWidth={62}
 				label={<Label
 					text="Display name"
-					description={"The name we\u2019ll use when sending emails from your app."}/>
+					description={'The name we\u2019ll use when sending emails from your app.'}/>
 				}
 				input={<TextInput
 					value={fields.sender_display_name}
@@ -218,7 +218,7 @@ export default class HostingSettings extends DashboardView {
 			<Field
 				labelWidth={62}
 				label={<Label
-					text={"Custom “choose a new password” page"}
+					text={'Custom “choose a new password” page'}
 					//getSiteDomain() is required here and not for the other templates because this template is an erb file, as opposed to the others which are html.
 					description={<span>This page will be loaded when users click on a reset password link. <a href={getSiteDomain() + '/apps/choose_password'} download="choose_password.html">Download the template</a>.</span>} />
 				}
@@ -230,7 +230,7 @@ export default class HostingSettings extends DashboardView {
 			<Field
 				labelWidth={62}
 				label={<Label
-					text={"Custom “password changed” page"}
+					text={'Custom “password changed” page'}
 					description={<span>This page will be loaded when users successfully change their password. <a href="/apps/password_reset_success.html" download="password_updated.html">Download the template</a>.</span>} />
 				}
 				input={<TextInput
@@ -241,7 +241,7 @@ export default class HostingSettings extends DashboardView {
 			<Field
 				labelWidth={62}
 				label={<Label
-					text={"Custom “email verified” page"}
+					text={'Custom “email verified” page'}
 					description={<span>This page will be loaded when users verify their email address. <a href="/apps/verify_email_success.html" download="email_verification.html">Download the template</a>.</span>} />
 				}
 				input={<TextInput
@@ -252,7 +252,7 @@ export default class HostingSettings extends DashboardView {
 			<Field
 				labelWidth={62}
 				label={<Label
-					text={"Custom 404 page"}
+					text={'Custom 404 page'}
 					description={<span>This page will be loaded whenever users mistype the reset password or verify email links. <a href="/apps/invalid_link.html" download="invalid_link.html">Download the template</a>.</span>} />
 				}
 				input={<TextInput
@@ -268,7 +268,7 @@ export default class HostingSettings extends DashboardView {
 			<Field
 				labelWidth={62}
 				label={<Label
-					text={"Parse Frame URL"}
+					text={'Parse Frame URL'}
 					description={<span>Upload <a href="/apps/user_management" download="user_management.html">this file</a> to your server and tell us where you put it.</span>} />
 				}
 				input={<TextInput
