@@ -10,8 +10,8 @@ import installDevTools from 'immutable-devtools';
 import React           from 'react';
 import ReactDOM        from 'react-dom';
 import Dashboard       from './Dashboard';
-import "core-js/stable";
-import "regenerator-runtime/runtime";
+import 'core-js/stable';
+import 'regenerator-runtime/runtime';
 
 require('stylesheets/fonts.scss');
 require('graphql-playground-react/build/static/css/index.css')
