@@ -96,7 +96,7 @@ function WebhookStore(state, action) {
         return Parse._request(
           'PUT',
           'hooks/functions/' + action.functionName,
-          { __op: "Delete" },
+          { __op: 'Delete' },
           { useMasterKey: true }
         ).then(() => {
           return state.set('webhooks', state.get('webhooks').filter(existingHook => existingHook.functionName != action.functionName));
@@ -105,7 +105,7 @@ function WebhookStore(state, action) {
         return Parse._request(
           'PUT',
           'hooks/triggers/' + action.triggerClass + '/' + action.triggerName,
-          { __op: "Delete" },
+          { __op: 'Delete' },
           { useMasterKey: true }
         ).then(() => {
           return state.set('webhooks', state.get('webhooks').filter(existingHook =>
